Develop

+++ Crunch-App/Views/Cart/Views/CartView.swift
@@ -14,43 +14,61 @@ struct CartView: View {
     init(realmManager: RealmManager) {
         _viewModel = StateObject(wrappedValue: CartViewModel(realmManager: realmManager))
     }
-   
+    @State private var paymentView: Bool = false
+    @State private var realmItems: UsableCartItems = UsableCartItems(name: "", image: "", price: 0, descript: "")
     
     var body: some View {
-        VStack (alignment: .leading, spacing: 20) {
-            List {
-                ForEach(viewModel.useableCartItems, id: \.self) { items in
-                    CartItem(item: items)
-                        .padding(.bottom, 10)
+        NavigationStack {
+            VStack (alignment: .leading, spacing: 20) {
+                if viewModel.useableCartItems.isEmpty {
+                    Spacer()
+                    EmptyCartView()
+                    Spacer()
                 }
-                .onDelete(perform: viewModel.deleteItem)
+                else {
+                    List {
+                        ForEach(viewModel.useableCartItems, id: \.self) { items in
+                            
+                            CartItem(cartViewModel: viewModel,
+                                     item: items)
+                            
+                            .padding(.bottom, 10)
+                        }
+                        .onDelete(perform: viewModel.deleteItem)
+                    }
+                    .listStyle(.plain)
+                    .scrollIndicators(.hidden)
+                    .buttonStyle(.plain)
+                    Spacer()
+                    VStack {
+                        Group {
+                            Divider()
+                            CheckoutMenu(text: "Total:", value: $viewModel.total)
+                        }
+                        .padding(.bottom, 10)
+                        
+                        CustomButton(title: "Checkout") {
+                            self.paymentView = true
+                            Task {
+                                try await viewModel.submitOrder()
+                            }
+                            
+                        }
+                    }
+                    .padding(.horizontal)
+                    
+                }
             }
-            .listStyle(.plain)
-            .background(
-                Color("SecondaryColor")
-                    .ignoresSafeArea())
-            HStack {
-                Text("Total:")
-                Spacer()
-                Text("\(viewModel.total)")
-            }
-            .font(.title2)
-            .fontWeight(.black)
-            Button(action: {}, label: {
-                Text("Checkout")
-                    .font(.title3)
-                    .fontWeight(.bold)
-                    .foregroundStyle(.white)
-                    .frame(height: 55)
-                    .frame(maxWidth: .infinity)
-                    .background(RoundedRectangle(cornerRadius: 10)
-                        .fill(Color("LightBlue")))
+            .fullScreenCover(isPresented: $paymentView, content: {
+                ProcessPaymentView()
             })
+            .navigationTitle("My Cart")
+            .navigationBarTitleDisplayMode(.inline)
+            .padding(.bottom, 20)
+            .background(
+                Color(AppColors.primaryLightGray)
+                    .ignoresSafeArea())
         }
-        .navigationTitle("My Cart")
-        .navigationBarTitleDisplayMode(.inline)
-        .padding(.horizontal)
-        .padding(.bottom, 20)
     }
     
 }
@@ -69,16 +87,6 @@ struct EmptyCartView: View {
                 .foregroundStyle(Color("LightGray"))
                 .padding(.horizontal)
                 .padding(.bottom, 30)
-            Button(action: {}, label: {
-                Text("Go Home")
-                    .font(.title3)
-                    .fontWeight(.bold)
-                    .foregroundStyle(.white)
-                    .frame(height: 55)
-                    .frame(maxWidth: .infinity)
-                    .background(RoundedRectangle(cornerRadius: 10)
-                        .fill(Color("LightBlue")))
-            })
         }
         .padding(.horizontal)
     }
@@ -86,4 +94,50 @@ struct EmptyCartView: View {
 
 #Preview {
     CartView(realmManager: RealmManager())
+    //    ProcessPaymentView()
+}
+struct ProcessPaymentView: View {
+    @State private var show: Bool = false
+    var body: some View {
+        NavigationView {
+            ZStack {
+                Color.primaryLightGray
+                VStack(spacing: 30) {
+                    ProgressView()
+                        .controlSize(.extraLarge)
+                    Text("Loading...")
+                        .font(.custom(AppFonts.semibold, size: 26))
+                    Text("Processing your payment... Please wait as we handle your transaction.")
+                        .font(.custom(AppFonts.regular, size: 22))
+                        .multilineTextAlignment(.center)
+                }
+                .padding(.horizontal)
+                .fullScreenCover(isPresented: $show, content: {
+                    SuccessState(viewModel: CartViewModel(realmManager: RealmManager()), show: $show)
+                })
+                .onAppear {
+                    DispatchQueue.main.asyncAfter(deadline: .now() + 5) {
+                        self.show = true
+                    }
+                }.navigationBarBackButtonHidden()
+            }
+        }
+        .background(
+            Color.green)
+    }
+}
+struct CheckoutMenu: View {
+    var text: String
+    @Binding var value: Int
+    var body: some View {
+        HStack {
+            Text(text)
+                .foregroundStyle(Color("LightGray"))
+                .font(.system(size: 16))
+            Spacer()
+            Text("R\(value)")
+                .font(.system(size: 20))
+                .fontWeight(.bold)
+        }
+    }
 }
